added game_gui_handler.py which switches between views
