--- restart_game.py
@@ -1,6 +1,0 @@
-
-
-#
-# def restart():
-#     sdg = SelectDifficultiesGUI()
-#     sdg.handle_events()
